fix update

+++ server/core/tasks.py
@@ -75,7 +75,7 @@ def track_audio_from_url(stream_id):
         if match_score > 0.9 * score:
             match_all = np.where(correlation > match_score * 0.95)
             match_indices = match_all[0]  
-            match_times = list(set(map(lambda x: np.round(x / sample_rate), match_indices)))
+            match_times = list(set(map(lambda x: int(np.round(x / sample_rate)), match_indices)))
             
             for match_time in match_times:
                 if match_time < 1800:
